fix loop counter error
when calling a node, I was earlier creating a new copy of the message
data AFTER the computation of the node, during the forward call.

this caused an error because the existing copy was updated, and because
Python uses references, some extra information was incorrectly added to
the message, which caused a lot of confusion

now the message is updated before anything happens, so we should be
fine.