use #include instead of specific imports
okay, a good import/module system has many weird edge cases, the
simplest of which is recursive imports or importing only a specific
flow.

instead of imports, we now have a #include statement, available in fbp
or a as an expression in lisp. #include takes a filename string,
evaluates the entire file without running any flow, and then adds the
flows in that file to global namespace.

in essence, #include-ing a file is the same as copy-pasting all the
flows from that file, so beware of name clashes. recursive imports
should work, though.

there does exist a more elegant solution to this, but it'll take a while
to clear out the edge cases.