added wraparound logic to turtle
use the % operator to override the forward call.  this ensures that all
nodes end up within the world frame, so that nasty negative numbers
don't show up when comparing node similarity.

the wraparound is not 100% perfect (because partial lines are not drawn
correctly) but otherwise it does the expected things.